+++ bin/check-process.rb
@@ -177,7 +177,7 @@ class CheckProcess < Sensu::Plugin::Check::CLI
     if File.exist?(path)
       File.read(path).strip.to_i
     else
-      critical "Could not read pid file #{path}"
+      false
     end
   end
 
@@ -245,9 +245,12 @@ class CheckProcess < Sensu::Plugin::Check::CLI
   def run
     procs = acquire_procs
 
-    if config[:file_pid] && (file_pid = read_pid(config[:file_pid]))
-      procs.reject! { |p| p[:pid].to_i != file_pid }
+    file_pid = nil
+    if config[:file_pid]
+      file_pid = read_pid(config[:file_pid])
+      procs.select! { |p| p[:pid].to_i == file_pid }
     end
+
     procs.reject! { |p| p[:pid].to_i == $PROCESS_ID } unless config[:match_self]
     procs.reject! { |p| p[:pid].to_i == Process.ppid } unless config[:match_parent]
     procs.reject! { |p| p[:command] =~ /#{config[:exclude_pat]}/ } if config[:exclude_pat]
@@ -275,7 +278,8 @@ class CheckProcess < Sensu::Plugin::Check::CLI
     msg += "; esec > #{config[:esec_over]}" if config[:esec_over]
     msg += "; csec < #{config[:cpu_under]}" if config[:cpu_under]
     msg += "; csec > #{config[:cpu_over]}" if config[:cpu_over]
-    msg += "; pid #{config[:file_pid]}" if config[:file_pid]
+    msg += "; pid #{config[:file_pid]}" if file_pid
+    msg += "; Could not read pid file #{config[:file_pid]}" if file_pid == false
 
     if config[:metric]
       # #YELLOW